Add form styling, validation, and required inputs
I have added 'required' to the required inputs in the HTML, the only
input that is optional is the phone number. I have also added some
styling to help confirm whether the input is valid or not (patterns for
the phone number and email) and will add another pattern for the password
input field shortly. However, the invalid pseudoclass is displaying on
page load and I would like that to only be displayed once the user clicks
out of the input. Will need to fix this.