Fix form validation to only show after blur
Fixed the issue where the form validation would show on page load (i.e.
red borders on all the required fields at the start). Now, the red
invalid borders will only show once the user clicks on and then clicks
away from the field while it is still invalid. If it is valid, then the
border will update to green to show it is validated. Accomplished this
by adding a blank placeholder to each required input field and using the
:not(:placeholder-shown) pseudoclass. Also minor CSS styling changes.